Updates to Jupyter notebook of Mesa Flocking model to remove errors
All code for the Mesa Flocking example, including visualizations. When running the viz in Jupyter, any change or rerun of the code requires a new port or it will throw an error saying the server is already running. This code provides a workaround, opening the server at a new port each time the code is run.