Removed heading as a parameter for movement & start of data collection
Updates to the Boids-based model of fish collective behaviour to remove heading/alignment as a parameter influencing agents' movement to allow polarization to become emergent. It can then also be used to measure cohesion. Therefore, have started to build in a datacollector of agent headings to report alongside the model.