Now correctly measures polarization using stdev of heading
Updated the function used in the data collector to correctly calculate standard deviation of heading as a metric for polarization. Heading (given in Cartesian x,y coordinates) is computed in radians and then the standard deviation is taken. Ideally, I'd like this metric to range from 0-1 for comparability across model runs, but that's later Sophie's problem.