Improved documentation, identified problem
Improved the docmumentation for the new cohere method and identified that the issue is the 'speed' variable in the Boid class. This is a scalar value and is positive - all of the boids go to the bottom right. When changed to negative, all of the boids go to the upper left.