Removed speed parameter, added match velocity
The speed parameter of the Agent class was always positive, leading the Boids to always head for the lower-right corner. I removed the speed parameter and added velocity (the magnitude/Euclidean distance of the heading vector) and then changed the function that returned neighbors' heading (match_heading) to return the neighbors' velocity and have the agents match that velocity. It's still not perfect, but at least the agents appear to be moving in all directions.