Calculating avoidance not separation and new randomization
Changed references to 'separation' to now be called 'avoidance' to describe the collision avoidance that the model needs. The goal is to stop the bouncing repulsions that are happening, causing boids to spring off in a different direction if they get too close. 

Also changed the heading randomization to be simpler.