Udated the shoal model to match the work done on boids_cohere_fix
The shoal model includes the function for standard deviation of heading that is used in the data collector and the graph of those data. This file has been updated to match the agent and model parameters developed in boids_cohere_fix.py, in an attempt to stop the boids from always moving to one corner. The changes include each agent heading towards the true center of mass of the group, and a new calculation of matching velocity (including the magnitude/speed) instead of matching heading (a scalar value without magnitude).