Fixed boids visualisation
The problem was that the file simple_continuous_canvas.js was trying to append the boids visualisation canvas to the HTML element with ID "gridwrapper". I guess after updating Mesa that element was no longer present on the page. The other two visualisations mount under the HTML with ID "elements" so I just add to that as well.