Renamed model_run.py to ichec_run.py, wasn't able to find shoal_model
Ran into an issue where the model_run.py (now ichec_run.py) couldn't import `shoal_model`. It was not in the correct folder so we moved ichec_run.py into the root of the project (so it sits next to shoal_model.py). This is the easy way to get it to work.