Created new ICHEC model run scripts for each of the parameters I vary
These scripts are based on the data_sensitivity.py script that I run locally. The difference here is that they're broken up into the different types so I can run each a set number of times.

The question is - what's the balance between the number of model runs inside each script and how many times I call the script on the taskfarm on the cluster? That seems like a bit of a tricky question to figure out, beyond trial and error. 

Also, hopefully, the data output will be as simple as it was for the test scripts, but somehow I doubt that will be the case....