Ok that was too many runs in each task, so now trying 300 x 3 x 200
Just for s & g, not expecting it to work. I do need to figure out if I can up the amount of computing power I can access here AND if running this multiple times will provide a similar distribution as a once-off, so I can get to 100k or 1 million runs.